requirements.txt
langchain>=0.3.0
langchain-groq>=0.2.0
python-dotenv>=1.0.0
streamlit>=1.30.0
